fold whole blocks in the any-length codec (#2)
* implement core error trait for the error types

* return u8 from the fixed width encoders

* fold whole blocks in the any-length codec

* update readme and benchmarks for the fold

* point repository url at spool-labs base58

* import encoded length constants in the backends

* size decoded_len for runs of leading ones

* reject decodings past the variable length limit

* run fmt clippy and tests in ci

* add fuzz targets for decode and roundtrip

* fold each block in place over a window

* record the in place fold numbers

* lay decoded bytes out without a staging buffer

* add an alloc feature for longer input

* record the zen 5 and long input numbers

* sum fold columns eight to a register

* pin a path when fuzzing on x86

* keep clippy and the msrv happy in ci

* settle the lints the x86 target reaches

* bump to 0.2.0 for the breaking changes

* adjust permissions

---------

+++ tests/decode64_readers.rs
@@ -49,7 +49,7 @@ fn wide_reader_matches_four_lane() {
     }
     for input in corpus() {
         let mut text = [0u8; MAX_ENCODED_64];
-        let length = encode_64(&input, &mut text);
+        let length = encode_64(&input, &mut text) as usize;
         let encoded = &text[..length];
 
         let mut wide_limbs = [0u32; 18];
